chore: rename project to PhishSOC across code and docs
Closes #17 modulo the deferred Cloudflare resource renames.

Replaces the Agentic-Inbox-era names from main and the PhishPilot
branding shipped in #21 with PhishSOC. PhishSOC speaks the buyer's
vocabulary directly (SOC == Security Operations Center) where Pilot
was generic AI-assistant framing in a crowded namespace.

User-visible name strings:
- README.md heading + intro + screenshot alt text
- hub/README.md project description
- package.json name + cloudflare label
- package-lock.json name (top-level + root package)
- app/root.tsx <title>
- app/routes/home.tsx meta title
- workers/mcp/index.ts MCP server name

Brand identity from #21:
- app/components/phishpilot/ -> app/components/phishsoc/ (5 imports updated)
- Logo wordmark "Phish/Pilot" -> "Phish/SOC" + comment
- index.css comments
- useUIStore STORAGE_KEY "phishpilot-ui" -> "phishsoc-ui" + comment
- public/theme-boot.js localStorage key (must match useUIStore)

Localstorage migration: changing the storage key resets persisted
theme/hue prefs for any preview users. SSR defaults (dark / Rust)
take over so first paint is unchanged; only manual customizations
are lost. Acceptable at POC stage.

Intentionally untouched (per #17 deferral):
- wrangler.jsonc worker name + R2 bucket_name / preview_bucket_name
  (worker rename = new deploy; R2 not in-place renameable)
- README.md R2 bucket setup instructions (match deferred bucket name)
- README.md Deploy-to-Cloudflare button URL (still upstream
  cloudflare/agentic-inbox; moves with the GitHub repo rename)
- pp- CSS class prefix (internal-only styling; rename is pure churn)
