+++ app/components/phishsoc/Sparkline.tsx
