+++ app/components/phishsoc/verdict.ts
